Started work on movement
Movement is locked on the X and Y axis.

Essentially 2D game but with 3D graphics and camera.

Movement is getting there but still not happy with it. Atm I was able to make it so that the player can steer their character in the air. However at the cost of no momentum effecting the player when the "forward" key is released. Which sucks lmao