Character Movement Rework
Today I decided that the movement that I had created last time was sub par and I was not a big fan, as I wrote in my previous commit. Therefore I decided to completely delete it and re-do it from scratch.

Much of my gripe had to do with Unity's own Character Controller (which handles gravity and other things). So I removed that component as well and got to work.

Here I wrote my own gravity and physics, allowing more precise tweaking and fixing. I am still at a stand still with the players force and velocity not continuing in the air, but alas.

I also added wall sliding and wall jumping!